Remove guard against no students for category
Every table will now be rendered even if there are no
students for that table.

In the case of j9, that means that tables Cook Island Maori,
Fijian, Niuean, etc. will all be rendered with 0 values if
there are no students for those tables in the moe file being
rendered